Break the travis build into three parts

+++ scripts/travis_script.sh
@@ -9,44 +9,53 @@ sh -e /etc/init.d/xvfb start || true
 
 export PATH="$HOME/miniconda/bin:$PATH"
 
-# Install and enable the server extension
-pip install -v .
-jupyter serverextension enable --py jupyterlab
+if [[ $GROUP == tests ]]; then
 
-npm run clean
-npm run build
-npm test
-npm run test:coverage
+    # Test against a clean npm build
+    npm run clean
+    npm run build:all
+    npm test
+
+    # Run the python tests
+    pushd jupyterlab
+    nosetests
+    popd
+
+    # Make sure we have CSS that can be converted with postcss
+    npm install -g postcss-cli
+    postcss jupyterlab/build/*.css > /dev/null
+
+    # Make sure we can start and kill the lab server
+    jupyter lab --no-browser &
+    TASK_PID=$!
+    # Make sure the task is running
+    ps -p $TASK_PID || exit 1
+    sleep 5
+    kill $TASK_PID
+    wait $TASK_PID
+fi
 
 
-# Run the python tests
-npm run build:serverextension
-python setup.py build
-pushd jupyterlab
-nosetests
-popd
-npm run build:examples
-npm run docs
-cp jupyter-plugins-demo.gif docs
+if [[ $GROUP == coverage_and_docs ]]; then
+    # Run the coverage check.
+    npm run test:coverage
 
-# Make sure we have CSS that can be converted with postcss
-npm install -g postcss-cli
-postcss jupyterlab/build/*.css > /dev/null
+    # Build the docs
+    npm run docs
+    cp jupyter-plugins-demo.gif docs
 
-# Verify docs build
-pushd tutorial
-conda env create -n test_docs -f environment.yml
-source activate test_docs
-make linkcheck
-make html
-source deactivate
-popd
+    # Verify docs build
+    pushd tutorial
+    conda env create -n test_docs -f environment.yml
+    source activate test_docs
+    make linkcheck
+    make html
+    source deactivate
+    popd
+fi
 
-# Make sure we can start and kill the lab server
-jupyter lab --no-browser &
-TASK_PID=$!
-# Make sure the task is running
-ps -p $TASK_PID || exit 1
-sleep 5
-kill $TASK_PID
-wait $TASK_PID
+
+if [[ $GROUP == examples ]]; then
+    # Make sure the examples build
+    npm run build:examples
+fi
